safari-534.53-branch/LayoutTests/storage/indexeddb/mozilla/create-objectstore-basics-expected.txt

Test IndexedDB's creating object store and updating properties

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = window.indexedDB || window.webkitIndexedDB || window.mozIndexedDB;
PASS indexedDB == null is false
IDBDatabaseException = window.IDBDatabaseException || window.webkitIDBDatabaseException;
PASS IDBDatabaseException == null is false
IDBTransaction = window.IDBTransaction || window.webkitIDBTransaction;
PASS IDBTransaction == null is false
indexedDB.open(name, description)
db = event.target.result
request = db.setVersion('1')
Deleted all object stores.
objectStore = db.createObjectStore(info.name, info.options);
PASS objectStore.name is info.name
PASS objectStore.indexNames.length is 0
PASS event.target.transaction.db is db
PASS event.target.transaction.readyState is IDBTransaction.LOADING
PASS event.target.transaction.mode is IDBTransaction.VERSION_CHANGE
objectStore = db.createObjectStore(info.name, info.options);
PASS objectStore.name is info.name
PASS objectStore.indexNames.length is 0
PASS event.target.transaction.db is db
PASS event.target.transaction.readyState is IDBTransaction.LOADING
PASS event.target.transaction.mode is IDBTransaction.VERSION_CHANGE
objectStore = db.createObjectStore(info.name, info.options);
PASS objectStore.name is info.name
PASS objectStore.indexNames.length is 0
PASS event.target.transaction.db is db
PASS event.target.transaction.readyState is IDBTransaction.LOADING
PASS event.target.transaction.mode is IDBTransaction.VERSION_CHANGE
objectStore = db.createObjectStore(info.name, info.options);
PASS objectStore.name is info.name
PASS objectStore.indexNames.length is 0
PASS event.target.transaction.db is db
PASS event.target.transaction.readyState is IDBTransaction.LOADING
PASS event.target.transaction.mode is IDBTransaction.VERSION_CHANGE
objectStore = db.createObjectStore(info.name, info.options);
PASS objectStore.name is info.name
PASS objectStore.indexNames.length is 0
PASS event.target.transaction.db is db
PASS event.target.transaction.readyState is IDBTransaction.LOADING
PASS event.target.transaction.mode is IDBTransaction.VERSION_CHANGE
objectStore = db.createObjectStore(info.name, info.options);
PASS objectStore.name is info.name
PASS objectStore.keyPath is info.options.keyPath
PASS objectStore.indexNames.length is 0
PASS event.target.transaction.db is db
PASS event.target.transaction.readyState is IDBTransaction.LOADING
PASS event.target.transaction.mode is IDBTransaction.VERSION_CHANGE
objectStore = db.createObjectStore(info.name, info.options);
PASS objectStore.name is info.name
PASS objectStore.keyPath is info.options.keyPath
PASS objectStore.indexNames.length is 0
PASS event.target.transaction.db is db
PASS event.target.transaction.readyState is IDBTransaction.LOADING
PASS event.target.transaction.mode is IDBTransaction.VERSION_CHANGE
objectStore = db.createObjectStore(info.name, info.options);
PASS objectStore.name is info.name
PASS objectStore.keyPath is info.options.keyPath
PASS objectStore.indexNames.length is 0
PASS event.target.transaction.db is db
PASS event.target.transaction.readyState is IDBTransaction.LOADING
PASS event.target.transaction.mode is IDBTransaction.VERSION_CHANGE
PASS successfullyParsed is true

TEST COMPLETE
